Add files via upload
Script used to analyze data recorded while measuring ultrasound beam profile in WELCOME - LLN.

Script not used after 4th of June 2018 --> computation of exposure indices not valid anymore. We can't use the whole signal to compute the PII because of high variations in the burst recorded while the positions were adjacent (pattern with lines, parallel to the beam propagation). All the analysis have to be re-performed with another script.